pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aurenet"
version = "0.1.0"
description = "RGB Keyboard Monitoring with CheckMK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openrgb-python>=0.3.6",
    "evdev>=1.7.1",
    "requests>=2.33.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.pytest.ini_options]
testpaths = ["aurenet/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=aurenet --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
